several fixes (including osx) merge with upstream some refactoring
