+++ assets/webconfig/content/generalconf.html
@@ -8,7 +8,7 @@
 			<hr>
 			<div class="col-lg-12">
 				<div id='editor_container'></div>
-				<button id='btn_submit'>Submit (console.log)</button>
+				<button id='btn_submit' class="btn btn-danger">Submit - currently will destroy your config!</button>
 			</div>
 		</div>
 	</div>
