+++ include/utils/FileUtils.h
@@ -3,6 +3,6 @@
 namespace FileUtils {
 
 std::string getBaseName( std::string sourceFile);
-std::string command_exec(const char* cmd);
+
 
 };
