+++ include/utils/Process.h
@@ -1,0 +1,8 @@
+#include <string>
+
+namespace Process {
+	
+void restartHyperion(bool asNewProcess=false); 
+std::string command_exec(const char* cmd);
+
+};
