+++ libsrc/utils/FileUtils.cpp
@@ -1,10 +1,5 @@
 #include <utils/FileUtils.h>
 
-#include <cstdio>
-#include <iostream>
-#include <memory>
-#include <stdexcept>
-
 #include <QFileInfo>
 
 namespace FileUtils {
@@ -15,21 +10,5 @@ std::string getBaseName( std::string sourceFile)
 	return fi.fileName().toStdString();
 }
  
-std::string command_exec(const char* cmd)
-{
-	char buffer[128];
-	std::string result = "";
-	std::shared_ptr<FILE> pipe(popen(cmd, "r"), pclose);
-	if (pipe) 
-	{
-	while (!feof(pipe.get()))
-	{
-		if (fgets(buffer, 128, pipe.get()) != NULL)
-			result += buffer;
-	}
-	}
-	
-	return result;
-}
 
 };
